+++ src/index.js
@@ -1,6 +1,6 @@
 export { parse } from './parse.js';
 export { merge } from './merge.js';
-export { generate } from './generate.js';
+export { generate, generateNode } from './generate.js';
 export { locate } from './locate.js';
 export { render } from './render.js';
 export { DEFAULT_CONTEXT } from './constants.js';